+++ instrumentation/runner/src/main/kotlin/de/mannodermaus/junit5/internal/runners/AndroidJUnit5.kt
@@ -1,10 +1,7 @@
 package de.mannodermaus.junit5.internal.runners
 
 import android.annotation.SuppressLint
-import android.util.Log
 import androidx.annotation.VisibleForTesting
-import de.mannodermaus.junit5.internal.LOG_TAG
-import de.mannodermaus.junit5.internal.LibcoreAccess
 import de.mannodermaus.junit5.internal.runners.notification.ParallelRunNotifier
 import org.junit.platform.engine.discovery.MethodSelector
 import org.junit.platform.launcher.core.LauncherFactory
@@ -44,12 +41,25 @@ internal class AndroidJUnit5(
     private fun generateTestTree(params: AndroidJUnit5RunnerParams): AndroidJUnitPlatformTestTree {
         val selectors = params.createSelectors(testClass)
         val isIsolatedMethodRun = selectors.size == 1 && selectors.first() is MethodSelector
+        val isUsingOrchestrator = params.isUsingOrchestrator
         val request = params.createDiscoveryRequest(selectors)
+
+        // Validate if run can be executed
+        if (isUsingOrchestrator && params.isParallelExecutionEnabled) {
+            throw RuntimeException(
+                """
+                    Running tests with the Android Test Orchestrator does not work with parallel tests,
+                    since some information must be retained across parallel test execution,
+                    and the isolated nature of the Android Test Orchestrator thwarts these efforts.
+                    Please disable either setting and try again.
+                """.trimIndent(),
+            )
+        }
 
         return AndroidJUnitPlatformTestTree(
             testPlan = launcher.discover(request),
             testClass = testClass,
-            isIsolatedMethodRun = isIsolatedMethodRun,
+            needLegacyFormat = isIsolatedMethodRun || isUsingOrchestrator,
             isParallelExecutionEnabled = params.isParallelExecutionEnabled,
         )
     }