+++ instrumentation/CHANGELOG.md
@@ -8,6 +8,7 @@ Change Log
 - Fix invalid naming of dynamic tests when executing only a singular test method from the IDE (#317, #339)
 - Prevent test methods incorrectly defined as Kotlin top-level functions from messing up Android's internal test counting, causing issues like "Expected N+1 tests, received N" (#316)
 - Prevent test classes ignored by a tag from being considered for test execution, causing issues like "Expected N+1 tests, received N" (#298)
+- Improve integration with Android Test Orchestrator and remove the need for `@UseTechnicalNames` (#337)
 
 ## 1.4.0 (2023-11-05)